Also use full paths for configuration files
Instead of a domain name, a complete path for a configuration file can
now be passed to the functions of the conf module.
If a name as $config is passed without an extension, it is assumed that
the name is a domain name. The corresponding file is then searched for
as usual under $__conf_root/${config}.conf.